backend/requirements.txt
appwrite==10.0.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==37.1.0
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.8.2
